app/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.5.3
sentence-transformers==2.5.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
numpy==1.26.3
supabase==2.3.4